+++ pubsub/gcloud/aio/pubsub/__init__.py
@@ -4,7 +4,7 @@ __version__ = get_distribution('gcloud-aio-pubsub').version
 from gcloud.aio.pubsub.publisher_client import PublisherClient
 from gcloud.aio.pubsub.subscriber_client import SubscriberClient
 from gcloud.aio.pubsub.utils import PubsubMessage
-from gcloud.aio.pubsub.subscriber_message import FlowControl
+from gcloud.aio.pubsub.subscriber_client import FlowControl
 from gcloud.aio.pubsub.subscriber_message import SubscriberMessage
