+++ pubsub/gcloud/aio/pubsub/subscriber_client.py
@@ -1,13 +1,14 @@
 from gcloud.aio.auth import BUILD_GCLOUD_REST  # pylint: disable=no-name-in-module
 
 if BUILD_GCLOUD_REST:
+    class FlowControl:
+        def __init__(self, **kwargs) -> None:
+            raise NotImplementedError('this class is only implemented in aio')
+
     class SubscriberClient:
         def __init__(self, **kwargs) -> None:
             raise NotImplementedError('this class is only implemented in aio')
 
-    class FlowControl:
-        def __init__(self, **kwargs) -> None:
-            raise NotImplementedError('this class is only implemented in aio')
 else:
     import asyncio
     import concurrent.futures
@@ -21,8 +22,20 @@ else:
     from google.cloud.pubsub_v1.subscriber.message import Message
 
     from .subscriber_message import SubscriberMessage
-    from .subscriber_message import FlowControl
     from .utils import convert_google_future_to_concurrent_future
+
+
+    class FlowControl:
+        def __init__(self,
+                     max_bytes: int = 100 * 1024 * 1024,
+                     max_messages: int = 1000,
+                     max_lease_duration: int = 1 * 60 * 60,
+                     max_duration_per_lease_extension: int = 0) -> None:
+            self.max_bytes = max_bytes
+            self.max_messages = max_messages
+            self.max_lease_duration = max_lease_duration
+            self.max_duration_per_lease_extension = (
+                max_duration_per_lease_extension)
 
 
     class SubscriberClient:
